Update progress range from 1-100 to 0-1 float range in onProgressChanged

+++ ebookreader/src/main/java/com/vipulog/ebookreader/EbookReaderEventListener.kt
@@ -3,6 +3,6 @@ package com.vipulog.ebookreader
 interface EbookReaderEventListener {
     fun onBookLoaded(bookMetaData: BookMetaData)
     fun onBookLoadFailed(error: String)
-    fun onProgressChanged(progress: Int, currentTocItem: TocItem?)
+    fun onProgressChanged(progress: Float, currentTocItem: TocItem?)
     fun onTextSelectionModeChange(mode: Boolean)
 }
